[License] Add official licenses to PDELab Howto
The PDELab howto is now officially licensed, using the BSD 3-clause license for the software portion and Creative Commons Attribution-ShareAlike 3.0 for the Howto document.

The actual licenses are in external files BSD and CC-BY-SA. I had to copy all copyright holders into the copyright notice at the top of the BSD license, as otherwise, the license apparently cannot officially be called BSD 3-clause.

The patch also adds an explicit copyright notice to the Howto-documents, both as commentary in the TeX source and on the front pages of the respective PDF output files.

Imported from SVN: r616 at trunk