[dune-course-2015] Work around a weird problem when compiling with latex-beamer for the Heidelberg Dune Course 2015:
ERROR: Undefined control sequence.

--- TeX said ---
l.983 \paragraph
                {Vertex-centered Finite Volumes}
--- HELP ---
TeX encountered an unknown command name. You probably misspelled the
name. If this message occurs when a LaTeX command is being processed,
the command is probably in the wrong place---for example, the error
can be produced by an \item command that's not inside a list-making
environment. The error can also be caused by a missing \documentclass
command.