[Cleanup][NavierStokes] merge implementation of ...
... boundary and initial conditions from "cgstokes_initial.hh" and
"sproblemA.hh" to "navierstokes_initial.hh".

Rename the parameter class "HagenPoiseuilleVelocity" to
"HagenPoiseuilleVelocityBox". This class is used in the HY2, HA2 options
of "cgstokes.cc" and in "dgstokes.cc" .

Make the template specialization of "HagenPoiseuilleVelocity" in three
dimensions to separate class "HagenPoiseuilleVelocitySpherical". This
class is used in the HU3 option of "cgstokes.cc" instead.