[bugfix,release] Correctly instantiate linear problem for reentrant corner.
Either the order of the constructor parameters must have changed or this
has never compiled. Anyway, previously a non-existent constructor was called
which did not compile.

This commit changes the order of the constructor parameter passed to StationaryLinearProblemSolver as expected there.

The problem occured on releases/2.4. Note that I did not test the compilation
of this patch on master as compiling the howto takes ages :/